Fixed audio and video syncing.

+++ brainfrick.h
@@ -7,7 +7,9 @@
 #include <stdbool.h>
 #include <stdio.h>
 #include <math.h>
+#include <time.h>
 #include <sys/timeb.h>
+#include <unistd.h>
 #ifndef _WIN32
     #include <ncurses.h>
 #else
@@ -21,7 +23,7 @@ struct bf_t {
     size_t pos;
 };
 
-void interpret_code(const char*, bool, size_t);
+void interpret_code(const char*, bool, double);
 void init_bf(void);
 void end_bf(void);
 void set_frame_size(size_t);
